requirements.txt
discord.py
discord-py-interactions
python-dotenv
openai
bs4
urllib3